Attempt to fix compilation on win32

+++ src/uniroot.cpp
@@ -6,6 +6,6 @@
 double test_uniroot(Rcpp::Function f, double min, double max) {
   using namespace plant::util;
   const double tol = 1e-6;
-  const int max_iterations = 100;
+  const size_t max_iterations = 100;
   return uniroot(RFunctionWrapper(f), min, max, tol, max_iterations);
 }
